Move mui ThemeProvider to root
