+++ app/src/components/shared/TimeSeriesStatistics.tsx
@@ -4,7 +4,7 @@ import { getCurrentLanguageInformation } from "../../utils/utils";
 import { MuiPickersUtilsProvider } from "@material-ui/pickers";
 import DateFnsUtils from "@date-io/date-fns";
 import { DatePicker } from "@material-ui/pickers";
-import { createTheme, ThemeProvider } from "@material-ui/core";
+import { createTheme } from "@material-ui/core";
 import { Field, Formik } from "formik";
 import BarChart from "./BarChart";
 import {
@@ -322,49 +322,45 @@ const TimeSeriesStatistics = ({
 									{/* date picker for selecting start date of the statistic */}
 									<span>{t("STATISTICS.FROM") /* From */}</span>
 									<div className="chosen-container">
-										<ThemeProvider theme={theme}>
-											<DatePicker
-												name="fromDate"
-												style={datePickerStyle}
-												value={formik.values.fromDate}
-												placeholder={t(
-													"EVENTS.EVENTS.NEW.SOURCE.PLACEHOLDER.START_DATE"
-												)}
-												onChange={(value) =>
-													changeFrom(
-														value,
-														formik.setFieldValue,
-														formik.values.timeMode,
-														formik.values.toDate,
-														formik.values.dataResolution
-													)
-												}
-											/>
-										</ThemeProvider>
+										<DatePicker
+											name="fromDate"
+											style={datePickerStyle}
+											value={formik.values.fromDate}
+											placeholder={t(
+												"EVENTS.EVENTS.NEW.SOURCE.PLACEHOLDER.START_DATE"
+											)}
+											onChange={(value) =>
+												changeFrom(
+													value,
+													formik.setFieldValue,
+													formik.values.timeMode,
+													formik.values.toDate,
+													formik.values.dataResolution
+												)
+											}
+										/>
 									</div>
 
 									{/* date picker for selecting end date of the statistic */}
 									<span>{t("STATISTICS.TO") /* To */}</span>
 									<div className="chosen-container">
-										<ThemeProvider theme={theme}>
-											<DatePicker
-												name="toDate"
-												style={datePickerStyle}
-												value={formik.values.toDate}
-												placeholder={t(
-													"EVENTS.EVENTS.NEW.SOURCE.PLACEHOLDER.END_DATE"
-												)}
-												onChange={(value) =>
-													changeTo(
-														value,
-														formik.setFieldValue,
-														formik.values.timeMode,
-														formik.values.fromDate,
-														formik.values.dataResolution
-													)
-												}
-											/>
-										</ThemeProvider>
+										<DatePicker
+											name="toDate"
+											style={datePickerStyle}
+											value={formik.values.toDate}
+											placeholder={t(
+												"EVENTS.EVENTS.NEW.SOURCE.PLACEHOLDER.END_DATE"
+											)}
+											onChange={(value) =>
+												changeTo(
+													value,
+													formik.setFieldValue,
+													formik.values.timeMode,
+													formik.values.fromDate,
+													formik.values.dataResolution
+												)
+											}
+										/>
 									</div>
 								</div>
